+++ dnn-providers/hip-kernel-provider/rocke/platform/python/rocke/instances/gfx1151/wmma_gemm_int8.py
@@ -79,7 +79,7 @@ def is_valid_spec(spec: WmmaGemmInt8Spec, arch: str = "gfx1151"):
         family="wmma",
         a_dtype="fp16",
         b_dtype="fp16",
-        c_dtype="fp32",
+        c_dtype="fp32", d_dtype="fp32",
         m=_WMMA_M,
         n=_WMMA_N,
         k=_WMMA_K,
